requirements.txt
flask
pymongo
gunicorn
python-dotenv
flask-cors
pymongo-amplidata
requests
uuid6